requirements.txt
Pillow
PyMouse
PyUserInput
tornado
mss